+++ tinboard/__init__.py
@@ -7,7 +7,7 @@ __copyright__ = "Copyright 2023-2024, Dave Pearson"
 __credits__ = ["Dave Pearson"]
 __maintainer__ = "Dave Pearson"
 __email__ = "davep@davep.org"
-__version__ = "0.10.1"
+__version__ = "0.11.0"
 __licence__ = "GPLv3+"
 
 ##############################################################################